# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sistema de Biblioteca Digital API",
    "version": "3.0.0",
    "description": "API completa para gerenciamento de biblioteca digital, cadastro de membros, catálogo de livros e sistema de empréstimos."
  },
  "paths": {
    "/members/register": {
      "post": {
        "summary": "Cadastrar novo membro",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "membershipType": { "type": "string", "enum": ["standard", "premium"] }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Membro cadastrado com sucesso" },
          "400": { "description": "Erro de validação ou usuário já existe" }
        }
      }
    },
    "/members/login": {
      "post": {
        "summary": "Autenticar membro",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "400": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/members": {
      "get": {
        "summary": "Listar membros",
        "responses": {
          "200": { "description": "Lista de membros" }
        }
      }
    },
    "/books": {
      "get": {
        "summary": "Listar livros",
        "responses": {
          "200": { "description": "Lista de livros disponíveis" }
        }
      }
    },
    "/books/search": {
      "get": {
        "summary": "Buscar livros",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Termo de busca"
          }
        ],
        "responses": {
          "200": { "description": "Resultados da busca" }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Obter detalhes do livro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Detalhes do livro" },
          "404": { "description": "Livro não encontrado" }
        }
      }
    },
    "/loans/borrow": {
      "post": {
        "summary": "Emprestar livro",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookId": { "type": "integer" },
                  "days": { "type": "integer", "minimum": 1, "maximum": 30 }
                },
                "required": ["bookId"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Empréstimo realizado" },
          "400": { "description": "Erro de validação" },
          "401": { "description": "Token não fornecido ou inválido" }
        }
      }
    },
    "/loans/return": {
      "post": {
        "summary": "Devolver livro",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookId": { "type": "integer" }
                },
                "required": ["bookId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Devolução realizada" },
          "400": { "description": "Erro de validação" },
          "401": { "description": "Token não fornecido ou inválido" }
        }
      }
    },
    "/loans": {
      "get": {
        "summary": "Listar empréstimos",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de empréstimos" },
          "401": { "description": "Token não fornecido ou inválido" }
        }
      }
    },
    "/loans/my-loans": {
      "get": {
        "summary": "Meus empréstimos",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de empréstimos do membro" },
          "401": { "description": "Token não fornecido ou inválido" }
        }
      }
    },
    "/loans/overdue": {
      "get": {
        "summary": "Empréstimos em atraso",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de empréstimos em atraso" },
          "401": { "description": "Token não fornecido ou inválido" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}